# Make sure times are sorted in the `at-benchmark` histogram