--- a/Project.toml
+++ b/Project.toml
@@ -1,6 +1,6 @@
 name = "BenchmarkTools"
 uuid = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
-version = "1.1.3"
+version = "1.1.4"
 
 [deps]
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"